@@ -5,14 +5,14 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 
 android {
-  compileSdkVersion 27
+  compileSdkVersion 28
   defaultConfig {
-    applicationId "codeka.com.au.podcreep"
+    applicationId "au.com.codeka.podcreep"
     minSdkVersion 24
-    targetSdkVersion 27
+    targetSdkVersion 28
     versionCode 1
     versionName "1.0"
-    testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
   }
   buildTypes {
     release {
@@ -28,11 +28,11 @@ android {
 dependencies {
   implementation fileTree(dir: 'libs', include: ['*.jar'])
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-  implementation 'com.android.support:appcompat-v7:27.1.1'
-  implementation 'com.android.support:design:27.1.1'
-  implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-  implementation 'com.android.support:recyclerview-v7:27.1.1'
-  implementation 'com.android.support:support-media-compat:27.1.1'
+  implementation 'androidx.appcompat:appcompat:1.1.0-alpha01'
+  implementation 'com.google.android.material:material:1.1.0-alpha02'
+  implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+  implementation 'androidx.recyclerview:recyclerview:1.1.0-alpha01'
+  implementation 'androidx.media:media:1.1.0-alpha01'
   implementation 'com.squareup.moshi:moshi:1.5.0'
   implementation 'com.squareup.moshi:moshi-kotlin:1.5.0'
   implementation 'com.squareup.picasso:picasso:2.71828'
