@@ -1,7 +1,7 @@
 package au.com.codeka.podcreep.app.podcasts.discover
 
 import android.content.Context
-import android.support.v4.view.PagerAdapter
+import androidx.viewpager.widget.PagerAdapter
 import android.util.Log
 import android.view.View
 import android.view.ViewGroup
