@@ -1,7 +1,7 @@
 package au.com.codeka.podcreep.ui
 
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.view.ViewGroup
 import kotlin.reflect.KClass
 
