@@ -7,6 +7,6 @@ import au.com.codeka.podcreep.concurrency.TaskRunner
 class SearchTabLayout(
     context: Context,
     taskRunner: TaskRunner,
-    callback: DiscoverLayout.Callbacks)
+    callbacks: DiscoverLayout.Callbacks)
   : LinearLayout(context) {
 }
